fix allocation of stream indexes
Stream indexes were allocated using a global counter. This means that, if one
creates two logfile sequentially with (let's say) 10 and 20 streams each, then
the second set of streams would be numbered [10, 19].

This causes a problem for long-running processes, or for processes which would
create new logfiles often. The problem does not so much lie on the logging
process itself (in its case, the logging index does not matter much), but it
does matter on the reading side as the log streams are usually stored in an
array so that they can be addresses by their index. If there is too much holes,
that causes some problems.

This commit creates a Logfile class which takes care of the stream index
allocation per-file instead of globally, thus fixing the problem.