adapt for orogen backport the TypelibMarshaller from the rtt2.0 branch
The isse we are trying to fix here is that some types are not
marshalled as themselves, but as intermediate types. Until now,
there was a HUGE hack that was simply testing for type_name + _m,
as orogen was generating these types for intermediate types.

This does not work for purely opaque types, only for types that
contain opaques.

The proper fix is to have the type marshaller object tell us
what on-disk type it is using. That is what this patch is doing.